Added Create Date to Momento Map, changed the integers for NewTags to decimals.

+++ Source/Momntz.UI/Core/Binders/NewTagModelBinder.cs
@@ -9,10 +9,10 @@ namespace Momntz.UI.Core.Binders
         public object BindModel(ControllerContext controllerContext, ModelBindingContext bindingContext)
         {
             NewTag tag = new NewTag();
-            tag.Left = Convert.ToInt32(controllerContext.HttpContext.Request["left"]);
-            tag.Top = Convert.ToInt32(controllerContext.HttpContext.Request["top"]);
-            tag.Height = Convert.ToInt32(controllerContext.HttpContext.Request["height"]);
-            tag.Width = Convert.ToInt32(controllerContext.HttpContext.Request["width"]);
+            tag.Left = Convert.ToDecimal(controllerContext.HttpContext.Request["left"]);
+            tag.Top = Convert.ToDecimal(controllerContext.HttpContext.Request["top"]);
+            tag.Height = Convert.ToDecimal(controllerContext.HttpContext.Request["height"]);
+            tag.Width = Convert.ToDecimal(controllerContext.HttpContext.Request["width"]);
             tag.MomentoId = Convert.ToInt32(controllerContext.HttpContext.Request.QueryString["image_id"]);
             tag.Name = controllerContext.HttpContext.Request["name"];
